+++ routes/api.php
@@ -3,6 +3,7 @@
 use Illuminate\Support\Facades\Route;
 use Kami\Cocktail\Http\Controllers\AuthController;
 use Kami\Cocktail\Http\Controllers\UserController;
+use Kami\Cocktail\Http\Controllers\GlassController;
 use Kami\Cocktail\Http\Controllers\ImageController;
 use Kami\Cocktail\Http\Controllers\ShelfController;
 use Kami\Cocktail\Http\Controllers\HealthController;
@@ -62,7 +63,7 @@ Route::middleware('auth:sanctum')->group(function() {
         Route::get('/user-shelf', [CocktailController::class, 'userShelf'])->name('cocktails.user-shelf');
         Route::get('/user-favorites', [CocktailController::class, 'userFavorites'])->name('cocktails.user-favorites');
         Route::get('/{id}', [CocktailController::class, 'show'])->name('cocktails.show');
-        Route::post('/{id}/favorite', [CocktailController::class, 'favorite'])->name('cocktails.favorite');
+        Route::post('/{id}/toggle-favorite', [CocktailController::class, 'toggleFavorite'])->name('cocktails.favorite');
         Route::post('/', [CocktailController::class, 'store'])->name('cocktails.store');
         Route::delete('/{id}', [CocktailController::class, 'delete'])->name('cocktails.delete');
         Route::put('/{id}', [CocktailController::class, 'update'])->name('cocktails.update');
@@ -71,12 +72,18 @@ Route::middleware('auth:sanctum')->group(function() {
     Route::prefix('images')->group(function() {
         Route::get('/{id}', [ImageController::class, 'show']);
         Route::post('/', [ImageController::class, 'store']);
+        Route::post('/{id}', [ImageController::class, 'update']);
         Route::delete('/{id}', [ImageController::class, 'delete']);
     });
 
     Route::prefix('shopping-lists')->group(function() {
         Route::post('/batch', [ShoppingListController::class, 'batchStore']);
         Route::delete('/batch', [ShoppingListController::class, 'batchDelete']);
+    });
+
+    Route::prefix('glasses')->group(function() {
+        Route::get('/', [GlassController::class, 'index']);
+        Route::get('/{id}', [GlassController::class, 'show']);
     });
 
 });